# log: Add TimestampFormat and Timeformat.