--- a/log/value_test.go
+++ b/log/value_test.go
@@ -1,7 +1,9 @@
 package log_test
 
 import (
+	"encoding"
 	"fmt"
+	"reflect"
 	"testing"
 	"time"
 
@@ -34,7 +36,7 @@ func TestValueBinding(t *testing.T) {
 	if want, have := start.Add(time.Second), timestamp; want != have {
 		t.Errorf("output[1]: want %v, have %v", want, have)
 	}
-	if want, have := "value_test.go:29", fmt.Sprint(output[3]); want != have {
+	if want, have := "value_test.go:31", fmt.Sprint(output[3]); want != have {
 		t.Errorf("output[3]: want %s, have %s", want, have)
 	}
 
@@ -47,7 +49,7 @@ func TestValueBinding(t *testing.T) {
 	if want, have := start.Add(2*time.Second), timestamp; want != have {
 		t.Errorf("output[1]: want %v, have %v", want, have)
 	}
-	if want, have := "value_test.go:42", fmt.Sprint(output[3]); want != have {
+	if want, have := "value_test.go:44", fmt.Sprint(output[3]); want != have {
 		t.Errorf("output[3]: want %s, have %s", want, have)
 	}
 }
@@ -87,6 +89,43 @@ func TestValueBinding_loggingZeroKeyvals(t *testing.T) {
 	}
 	if want, have := start.Add(2*time.Second), timestamp; want != have {
 		t.Errorf("output[1]: want %v, have %v", want, have)
+	}
+}
+
+func TestTimestampFormat(t *testing.T) {
+	t.Parallel()
+
+	start := time.Date(2015, time.April, 25, 0, 0, 0, 0, time.UTC)
+	now := start
+	mocktime := func() time.Time {
+		now = now.Add(time.Second)
+		return now
+	}
+
+	tv := log.TimestampFormat(mocktime, time.RFC822)
+
+	if want, have := now.Add(time.Second).Format(time.RFC822), fmt.Sprint(tv()); want != have {
+		t.Errorf("wrong time format: want %v, have %v", want, have)
+	}
+
+	if want, have := now.Add(2*time.Second).Format(time.RFC822), fmt.Sprint(tv()); want != have {
+		t.Errorf("wrong time format: want %v, have %v", want, have)
+	}
+
+	mustMarshal := func(v interface{}) []byte {
+		b, err := v.(encoding.TextMarshaler).MarshalText()
+		if err != nil {
+			t.Fatal("error marshaling text:", err)
+		}
+		return b
+	}
+
+	if want, have := now.Add(3*time.Second).AppendFormat(nil, time.RFC822), mustMarshal(tv()); !reflect.DeepEqual(want, have) {
+		t.Errorf("wrong time format: want %s, have %s", want, have)
+	}
+
+	if want, have := now.Add(4*time.Second).AppendFormat(nil, time.RFC822), mustMarshal(tv()); !reflect.DeepEqual(want, have) {
+		t.Errorf("wrong time format: want %s, have %s", want, have)
 	}
 }