- TimeFormat provides a value that defers formating timestamps until
  serialization, which improves performance when a log event is not
  serialized.
- TimestampFormat is a convenience function to produce time Valuers
  with arbitrary time formats.
- DefaultTimestamp and DefaultTimestampUTC are now created using
  TimestampFormat.

Benchmarks from log/level showing the benefits:
```
name                                              old time/op    new time/op    delta
/TimeContext/Baseline/Nop-8                          679ns ± 0%     358ns ± 0%  -47.25%  (p=0.008 n=5+5)
/TimeContext/Baseline/Logfmt-8                      1.41µs ± 0%    1.49µs ± 0%   +5.67%  (p=0.008 n=5+5)
/TimeContext/Baseline/JSON-8                        2.87µs ± 0%    2.87µs ± 0%     ~     (p=0.611 n=5+5)
/TimeContext/DisallowedLevel/Nop-8                   680ns ± 0%     360ns ± 0%  -47.13%  (p=0.008 n=5+5)
/TimeContext/DisallowedLevel/Logfmt-8                680ns ± 0%     359ns ± 0%  -47.18%  (p=0.008 n=5+5)
/TimeContext/DisallowedLevel/JSON-8                  680ns ± 0%     360ns ± 0%  -47.04%  (p=0.008 n=5+5)
/TimeContext/AllowedLevel/Nop-8                      688ns ± 0%     364ns ± 0%  -47.15%  (p=0.016 n=4+5)
/TimeContext/AllowedLevel/Logfmt-8                  1.42µs ± 0%    1.50µs ± 0%   +5.38%  (p=0.008 n=5+5)
/TimeContext/AllowedLevel/JSON-8                    2.88µs ± 0%    2.88µs ± 0%   +0.32%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/Baseline/Nop-8              1.41µs ± 0%    1.07µs ± 1%  -24.43%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/Baseline/Logfmt-8           3.26µs ± 0%    3.37µs ± 0%   +3.55%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/Baseline/JSON-8             5.32µs ± 0%    5.34µs ± 0%   +0.53%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/DisallowedLevel/Nop-8       1.41µs ± 0%    1.06µs ± 0%  -24.49%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/DisallowedLevel/Logfmt-8    1.41µs ± 0%    1.06µs ± 0%  -24.53%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/DisallowedLevel/JSON-8      1.41µs ± 0%    1.07µs ± 1%  -24.37%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/AllowedLevel/Nop-8          1.42µs ± 0%    1.07µs ± 0%  -24.54%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/AllowedLevel/Logfmt-8       3.26µs ± 1%    3.38µs ± 0%   +3.81%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/AllowedLevel/JSON-8         5.35µs ± 0%    5.34µs ± 0%     ~     (p=0.238 n=5+5)

name                                              old alloc/op   new alloc/op   delta
/TimeContext/Baseline/Nop-8                           352B ± 0%      352B ± 0%     ~     (all equal)
/TimeContext/Baseline/Logfmt-8                        352B ± 0%      400B ± 0%  +13.64%  (p=0.008 n=5+5)
/TimeContext/Baseline/JSON-8                        1.26kB ± 0%    1.31kB ± 0%   +3.80%  (p=0.008 n=5+5)
/TimeContext/DisallowedLevel/Nop-8                    352B ± 0%      352B ± 0%     ~     (all equal)
/TimeContext/DisallowedLevel/Logfmt-8                 352B ± 0%      352B ± 0%     ~     (all equal)
/TimeContext/DisallowedLevel/JSON-8                   352B ± 0%      352B ± 0%     ~     (all equal)
/TimeContext/AllowedLevel/Nop-8                       352B ± 0%      352B ± 0%     ~     (all equal)
/TimeContext/AllowedLevel/Logfmt-8                    352B ± 0%      400B ± 0%  +13.64%  (p=0.008 n=5+5)
/TimeContext/AllowedLevel/JSON-8                    1.26kB ± 0%    1.31kB ± 0%   +3.80%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/Baseline/Nop-8                560B ± 0%      560B ± 0%     ~     (all equal)
/TimeCallerReqIDContext/Baseline/Logfmt-8             736B ± 0%      784B ± 0%   +6.52%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/Baseline/JSON-8             1.85kB ± 0%    1.90kB ± 0%   +2.60%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/DisallowedLevel/Nop-8         560B ± 0%      560B ± 0%     ~     (all equal)
/TimeCallerReqIDContext/DisallowedLevel/Logfmt-8      560B ± 0%      560B ± 0%     ~     (all equal)
/TimeCallerReqIDContext/DisallowedLevel/JSON-8        560B ± 0%      560B ± 0%     ~     (all equal)
/TimeCallerReqIDContext/AllowedLevel/Nop-8            560B ± 0%      560B ± 0%     ~     (all equal)
/TimeCallerReqIDContext/AllowedLevel/Logfmt-8         736B ± 0%      784B ± 0%   +6.52%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/AllowedLevel/JSON-8         1.85kB ± 0%    1.90kB ± 0%   +2.60%  (p=0.008 n=5+5)

name                                              old allocs/op  new allocs/op  delta
/TimeContext/Baseline/Nop-8                           8.00 ± 0%      7.00 ± 0%  -12.50%  (p=0.008 n=5+5)
/TimeContext/Baseline/Logfmt-8                        8.00 ± 0%      8.00 ± 0%     ~     (all equal)
/TimeContext/Baseline/JSON-8                          27.0 ± 0%      27.0 ± 0%     ~     (all equal)
/TimeContext/DisallowedLevel/Nop-8                    8.00 ± 0%      7.00 ± 0%  -12.50%  (p=0.008 n=5+5)
/TimeContext/DisallowedLevel/Logfmt-8                 8.00 ± 0%      7.00 ± 0%  -12.50%  (p=0.008 n=5+5)
/TimeContext/DisallowedLevel/JSON-8                   8.00 ± 0%      7.00 ± 0%  -12.50%  (p=0.008 n=5+5)
/TimeContext/AllowedLevel/Nop-8                       8.00 ± 0%      7.00 ± 0%  -12.50%  (p=0.008 n=5+5)
/TimeContext/AllowedLevel/Logfmt-8                    8.00 ± 0%      8.00 ± 0%     ~     (all equal)
/TimeContext/AllowedLevel/JSON-8                      27.0 ± 0%      27.0 ± 0%     ~     (all equal)
/TimeCallerReqIDContext/Baseline/Nop-8                9.00 ± 0%      8.00 ± 0%  -11.11%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/Baseline/Logfmt-8             14.0 ± 0%      14.0 ± 0%     ~     (all equal)
/TimeCallerReqIDContext/Baseline/JSON-8               38.0 ± 0%      38.0 ± 0%     ~     (all equal)
/TimeCallerReqIDContext/DisallowedLevel/Nop-8         9.00 ± 0%      8.00 ± 0%  -11.11%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/DisallowedLevel/Logfmt-8      9.00 ± 0%      8.00 ± 0%  -11.11%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/DisallowedLevel/JSON-8        9.00 ± 0%      8.00 ± 0%  -11.11%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/AllowedLevel/Nop-8            9.00 ± 0%      8.00 ± 0%  -11.11%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/AllowedLevel/Logfmt-8         14.0 ± 0%      14.0 ± 0%     ~     (all equal)
/TimeCallerReqIDContext/AllowedLevel/JSON-8           38.0 ± 0%      38.0 ± 0%     ~     (all equal)
```