@omeid 2017-04-17
One of the best traits of go-kit in general, and log package, is the minimal interface.

If `TimestampFormat` on it's own is enough; I think `TimeFormat` should be unexported so to keep the log package interface minimal and afford not committing to the specific implementation.