@ChrisHines 2017-04-19
Thanks for the input everyone. I'll make updates based on your feedback this evening.